# 8339387: ZGC: Synchronize medium page allocation