--- a/src/hotspot/share/gc/z/zObjectAllocator.cpp
+++ b/src/hotspot/share/gc/z/zObjectAllocator.cpp
@@ -25,6 +25,7 @@
 #include "gc/z/zGlobals.hpp"
 #include "gc/z/zHeap.inline.hpp"
 #include "gc/z/zHeuristics.hpp"
+#include "gc/z/zLock.inline.hpp"
 #include "gc/z/zObjectAllocator.hpp"
 #include "gc/z/zPage.inline.hpp"
 #include "gc/z/zPageTable.inline.hpp"
@@ -45,8 +46,9 @@ ZObjectAllocator::ZObjectAllocator(ZPageAge age)
     _use_per_cpu_shared_small_pages(ZHeuristics::use_per_cpu_shared_small_pages()),
     _used(0),
     _undone(0),
+    _shared_small_page(nullptr),
     _shared_medium_page(nullptr),
-    _shared_small_page(nullptr) {}
+    _medium_page_alloc_lock() {}
 
 ZPage** ZObjectAllocator::shared_small_page_addr() {
   return _use_per_cpu_shared_small_pages ? _shared_small_page.addr() : _shared_small_page.addr(0);
@@ -77,17 +79,20 @@ void ZObjectAllocator::undo_alloc_page(ZPage* page) {
   ZHeap::heap()->undo_alloc_page(page);
 }
 
+zaddress ZObjectAllocator::alloc_object_in_page_atomic(ZPage* page, size_t size) {
+  if (page == nullptr) {
+    return zaddress::null;
+  }
+  return page->alloc_object_atomic(size);
+}
+
 zaddress ZObjectAllocator::alloc_object_in_shared_page(ZPage** shared_page,
                                                        ZPageType page_type,
                                                        size_t page_size,
                                                        size_t size,
                                                        ZAllocationFlags flags) {
-  zaddress addr = zaddress::null;
   ZPage* page = Atomic::load_acquire(shared_page);
-
-  if (page != nullptr) {
-    addr = page->alloc_object_atomic(size);
-  }
+  zaddress addr = alloc_object_in_page_atomic(page, size);
 
   if (is_null(addr)) {
     // Allocate new page
@@ -126,6 +131,41 @@ zaddress ZObjectAllocator::alloc_object_in_shared_page(ZPage** shared_page,
   return addr;
 }
 
+zaddress ZObjectAllocator::alloc_object_in_medium_page(ZPageType page_type,
+                                                       size_t page_size,
+                                                       size_t size,
+                                                       ZAllocationFlags flags) {
+  ZPage** shared_page = _shared_medium_page.addr();
+  ZPage* page = Atomic::load_acquire(shared_page);
+
+  zaddress addr = alloc_object_in_page_atomic(page, size);
+
+  if (is_null(addr)) {
+    // When a new medium page is required, we synchronize the allocation
+    // of the new page using a lock. This is to avoid having multiple
+    // threads requesting a medium page from the page cache when we know
+    // only one of the will succeed in installing the page at this layer.
+    ZLocker<ZConditionLock> locker(&_medium_page_alloc_lock);
+
+    // When holding the lock we can't allow the page allocator to stall,
+    // which in the common case it won't. The page allocation is thus done
+    // in a non-blocking fashion and only if this fails we below (while not
+    // holding the lock) do the blocking page allocation.
+    ZAllocationFlags non_blocking_flags = flags;
+    non_blocking_flags.set_non_blocking();
+
+    addr = alloc_object_in_shared_page(_shared_medium_page.addr(), page_type, page_size, size, non_blocking_flags);
+  }
+
+  if (is_null(addr) && !flags.non_blocking()) {
+    // The above allocation attempts failed and this allocation should stall
+    // until memory is available. Redo the allocation with blocking enabled.
+    return alloc_object_in_shared_page(_shared_medium_page.addr(), page_type, page_size, size, flags);
+  }
+
+  return addr;
+}
+
 zaddress ZObjectAllocator::alloc_large_object(size_t size, ZAllocationFlags flags) {
   zaddress addr = zaddress::null;
 
@@ -141,7 +181,7 @@ zaddress ZObjectAllocator::alloc_large_object(size_t size, ZAllocationFlags flag
 }
 
 zaddress ZObjectAllocator::alloc_medium_object(size_t size, ZAllocationFlags flags) {
-  return alloc_object_in_shared_page(_shared_medium_page.addr(), ZPageType::medium, ZPageSizeMedium, size, flags);
+  return alloc_object_in_medium_page(ZPageType::medium, ZPageSizeMedium, size, flags);
 }
 
 zaddress ZObjectAllocator::alloc_small_object(size_t size, ZAllocationFlags flags) {